requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic[email]>=2.7.0
pydantic-settings>=2.2.1
python-multipart>=0.0.9
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
SQLAlchemy>=2.0.30
alembic>=1.13.2
psycopg[binary]>=3.1.19
httpx>=0.27.0
celery>=5.4.0
redis>=5.0.7
pdf2image>=1.17.0
Pillow>=10.4.0
groq>=0.9.0
